refactor - define single fixture responsability
